pyproject.toml
[project]
name = "quant-analyst-agent"
version = "0.1.0"
description = "Automated financial analysis agent - 777DDog"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.0.0",
    "certifi>=2026.1.4",
    "fastapi>=0.115.0",
    "fastmcp>=2.14.5",
    "lxml>=6.0.2",
    "matplotlib>=3.8.0",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.34.0",
    "yfinance>=1.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
